add `dev build`

+++ src/background/index.ts
@@ -10,11 +10,13 @@ import { ExtensionConfigHelper } from "@/share/storage";
 const logger = new CureLogger("bg/index");
 logger.log("Cure Cure ~\\(≧▽≦)/~");
 
-// #cure-tip 测试时应该保留下载的内容，方便后续快速测试 epub 打包
-// 安装时创建数据库
-chrome.runtime.onInstalled.addListener(() => {
-    CureBookPageDB.Instance.remove_all();
-});
+// #cure-test 测试时应该保留下载的内容，方便后续快速测试 epub、pdf 的生成
+if (__IS_DEV__) {
+    // 安装插件时删除清空建数据库
+    chrome.runtime.onInstalled.addListener(() => {
+        CureBookPageDB.Instance.remove_all();
+    });
+}
 
 // #region 让插件只在阅读界面启用
